Add StoneSpawner
물량 돌 생성기, 즉 fps 떨구는 장치를 설계함 약 10000개부터 조금씩 더뎌지며, 2000~4000까지는 정상적 작동됨.

+++ Assets/Script/SO/ProjectileSO.cs
@@ -4,7 +4,7 @@ using UnityEngine.UIElements;
 [CreateAssetMenu(fileName = "ProjectileSO", menuName = "Scriptable Objects/ProjectileSO")]
 public class ProjectileSO : ScriptableObject
 {
-    public int Speed;
-    public int Angle;
-    public int Mass;
+    public float Speed;
+    public float Angle;
+    public float Mass;
 }
